Get rid of BindleConnectionInfo duplication
Drop it from spin_publish and reuse it from spin_loader.
